Create Role test
